Restructure how events are handled
Actions are to the player what Events are to the user. The InputProcessor reads input and decides if that corresponds to an Action or an Event. Events are picked up by the EventProcessor and modifies metagame information. The StateProcessor cares about this information, and will change the game state depending on what it sees. Many processors had to be moved around, but the flow is now much more logical