Player needs to meet job requirements before switching
The player needs a certain amount of stat points in order to change. The player needs to be the correct race. The player needs to have mastered X number of skills from job Y in order to change. Added a Rogue class. Player starts as Soldier. Improve readability of SkillDirectoryProcessor. Added message feedback when trying to swtich jobs. Reduce duplicate code in WearableProcessor. Change structure of SkillDirectoryComponent, and updated codebase to reflect this.